requirements.txt
mistralai
chromadb
streamlit
pandas
pypdf
python-dotenv